+++ aeon/distances/_euclidean.py
@@ -2,6 +2,7 @@
 __author__ = ["chrisholder", "tonybagnall"]
 
 import numpy as np
+import xxlimited
 from numba import njit
 
 from aeon.distances._squared import _univariate_squared_distance, squared_distance
@@ -67,14 +68,18 @@ def _univariate_euclidean_distance(x: np.ndarray, y: np.ndarray) -> float:
 
 
 @njit(cache=True, fastmath=True)
-def euclidean_pairwise_distance(X: np.ndarray) -> np.ndarray:
+def euclidean_pairwise_distance(X: np.ndarray, y: np.ndarray = None) -> np.ndarray:
     """Compute the euclidean pairwise distance between a set of time series.
 
     Parameters
     ----------
     X: np.ndarray, of shape (n_instances, n_channels, n_timepoints) or
-            (n_instances, n_timepoints)
+            (n_instances, n_timepoints) or (n_timepoints,)
         A collection of time series instances.
+    y: np.ndarray, of shape (m_instances, m_channels, m_timepoints) or
+            (m_instances, m_timepoints) or (m_timepoints,), default=None
+        A collection of time series instances.
+
 
     Returns
     -------
@@ -84,7 +89,8 @@ def euclidean_pairwise_distance(X: np.ndarray) -> np.ndarray:
     Raises
     ------
     ValueError
-        If x and y are not 2D or 3D arrays.
+        If X is not 2D or 3D array when only passing X.
+        If X and y are not 1D, 2D or 3D arrays when passing both X and y.
 
     Examples
     --------
@@ -96,13 +102,47 @@ def euclidean_pairwise_distance(X: np.ndarray) -> np.ndarray:
            [ 5.29150262,  0.        ,  5.19615242],
            [10.44030651,  5.19615242,  0.        ]])
     """
-    if X.ndim == 3:
-        return _euclidean_pairwise_distance(X)
-    if X.ndim == 2:
-        _X = X.reshape((X.shape[0], 1, X.shape[1]))
-        return _euclidean_pairwise_distance(_X)
+    if y is None:
+        # To self
+        if X.ndim == 3:
+            return _euclidean_pairwise_distance(X)
+        if X.ndim == 2:
+            _X = X.reshape((X.shape[0], 1, X.shape[1]))
+            return _euclidean_pairwise_distance(_X)
+        raise ValueError("x and y must be 2D or 3D arrays")
+    elif y.ndim == X.ndim:
+        # Multiple to multiple
+        if y.ndim == 3 and X.ndim == 3:
+            return _euclidean_from_multiple_to_multiple_distance(X, y)
+        if y.ndim == 2 and X.ndim == 2:
+            _x = X.reshape((X.shape[0], 1, X.shape[1]))
+            _y = y.reshape((y.shape[0], 1, y.shape[1]))
+            return _euclidean_from_multiple_to_multiple_distance(_x, _y)
+        if y.ndim == 1 and X.ndim == 1:
+            _x = X.reshape((1, 1, X.shape[0]))
+            _y = y.reshape((1, 1, y.shape[0]))
+            return _euclidean_from_multiple_to_multiple_distance(_x, _y)
+        raise ValueError("x and y must be 1D, 2D, or 3D arrays")
+    else:
+        # Single to multiple
+        smallest_dim_ts = X if X.ndim <= y.ndim else y
+        largest_dim_ts = X if X.ndim > y.ndim else y
 
-    raise ValueError("x and y must be 2D or 3D arrays")
+        if largest_dim_ts.ndim == 3 and smallest_dim_ts.ndim == 2:
+            _x = smallest_dim_ts.reshape(
+                (smallest_dim_ts.shape[0], 1, smallest_dim_ts.shape[1])
+            )
+            return _euclidean_from_multiple_to_multiple_distance(
+                _x, largest_dim_ts
+            )
+        elif largest_dim_ts.ndim == 2 and smallest_dim_ts.ndim == 1:
+            _x = smallest_dim_ts.reshape((1, 1, smallest_dim_ts.shape[0]))
+            _y = largest_dim_ts.reshape(
+                (largest_dim_ts.shape[0], 1, largest_dim_ts.shape[1])
+            )
+            return _euclidean_from_multiple_to_multiple_distance(_x, _y)
+        else:
+            raise ValueError("x and y must be 2D or 3D arrays")
 
 
 @njit(cache=True, fastmath=True)
@@ -119,116 +159,8 @@ def _euclidean_pairwise_distance(X: np.ndarray) -> np.ndarray:
 
 
 @njit(cache=True, fastmath=True)
-def euclidean_from_single_to_multiple_distance(
-    x: np.ndarray, y: np.ndarray
-) -> np.ndarray:
-    """Compute the euclidean distance between a single time series and multiple.
-
-    Parameters
-    ----------
-    x: np.ndarray, (n_channels, n_timepoints) or (n_timepoints,)
-        Single time series.
-    y: np.ndarray, of shape (m_instances, m_channels, m_timepoints) or
-            (m_instances, m_timepoints)
-        A collection of time series instances.
-
-    Returns
-    -------
-    np.ndarray (n_instances)
-        euclidean distance between the collection of instances in y and the time
-        series x.
-
-    Raises
-    ------
-    ValueError
-        If x and y are not 2D or 3D arrays.
-
-    Examples
-    --------
-    >>> import numpy as np
-    >>> from aeon.distances import euclidean_from_single_to_multiple_distance
-    >>> x = np.array([[1, 2, 3, 6]])
-    >>> y = np.array([[[1, 2, 3, 4]],[[4, 5, 6, 3]], [[7, 8, 9, 3]]])
-    >>> euclidean_from_single_to_multiple_distance(x, y)
-    array([ 2.        ,  6.        , 10.81665383])
-    """
-    if y.ndim == 3 and x.ndim == 2:
-        return _euclidean_from_single_to_multiple_distance(x, y)
-    if y.ndim == 2 and x.ndim == 1:
-        _x = x.reshape((1, x.shape[0]))
-        _y = y.reshape((y.shape[0], 1, y.shape[1]))
-        return _euclidean_from_single_to_multiple_distance(_x, _y)
-    else:
-        raise ValueError("x and y must be 2D or 3D arrays")
-
-
-@njit(cache=True, fastmath=True)
-def _euclidean_from_single_to_multiple_distance(
-    x: np.ndarray, y: np.ndarray
-) -> np.ndarray:
-    n_instances = y.shape[0]
-    distances = np.zeros(n_instances)
-
-    for i in range(n_instances):
-        distances[i] = euclidean_distance(x, y[i])
-
-    return distances
-
-
-@njit(cache=True, fastmath=True)
-def euclidean_from_multiple_to_multiple_distance(
-    x: np.ndarray, y: np.ndarray
-) -> np.ndarray:
-    """Compute the euclidean distance between two sets of time series.
-
-    If x and y are the same then you should use euclidean_pairwise_distance.
-
-    Parameters
-    ----------
-    x: np.ndarray, of shape (n_instances, n_channels, n_timepoints) or
-            (n_instances, n_timepoints) or (n_timepoints,)
-        A collection of time series instances.
-    y: np.ndarray, of shape (m_instances, m_channels, m_timepoints) or
-            (m_instances, m_timepoints) or (m_timepoints,)
-        A collection of time series instances.
-
-    Returns
-    -------
-    np.ndarray (n_instances, m_instances)
-        euclidean distance between two collections of time series, x and y.
-
-    Raises
-    ------
-    ValueError
-        If x and y are not 2D or 3D arrays.
-
-    Examples
-    --------
-    >>> import numpy as np
-    >>> from aeon.distances import euclidean_from_multiple_to_multiple_distance
-    >>> x = np.array([[[1, 2, 3, 3]],[[4, 5, 6, 9]], [[7, 8, 9, 22]]])
-    >>> y = np.array([[[11, 12, 13, 2]],[[14, 15, 16, 1]], [[17, 18, 19, 10]]])
-    >>> euclidean_from_multiple_to_multiple_distance(x, y)
-    array([[17.34935157, 22.60530911, 28.58321186],
-           [14.        , 19.07878403, 22.53885534],
-           [21.16601049, 24.24871131, 21.07130751]])
-    """
-    if y.ndim == 3 and x.ndim == 3:
-        return _euclidean_from_multiple_to_multiple_distance(x, y)
-    if y.ndim == 2 and x.ndim == 2:
-        _x = x.reshape((x.shape[0], 1, x.shape[1]))
-        _y = y.reshape((y.shape[0], 1, y.shape[1]))
-        return _euclidean_from_multiple_to_multiple_distance(_x, _y)
-    if y.ndim == 1 and x.ndim == 1:
-        _x = x.reshape((1, 1, x.shape[0]))
-        _y = y.reshape((1, 1, y.shape[0]))
-        return _euclidean_from_multiple_to_multiple_distance(_x, _y)
-    raise ValueError("x and y must be 1D, 2D, or 3D arrays")
-
-
-@njit(cache=True, fastmath=True)
 def _euclidean_from_multiple_to_multiple_distance(
-    x: np.ndarray, y: np.ndarray
+        x: np.ndarray, y: np.ndarray
 ) -> np.ndarray:
     n_instances = x.shape[0]
     m_instances = y.shape[0]
