+++ aeon/distances/tests/test_new_distances.py
@@ -14,8 +14,6 @@ from aeon.distances import (
     dtw_from_single_to_multiple_distance,
     dtw_pairwise_distance,
     euclidean_distance,
-    euclidean_from_multiple_to_multiple_distance,
-    euclidean_from_single_to_multiple_distance,
     euclidean_pairwise_distance,
     squared_distance,
     squared_from_multiple_to_multiple_distance,
@@ -30,8 +28,8 @@ DISTANCES = [
         "name": "euclidean",
         "distance": euclidean_distance,
         "pairwise_distance": euclidean_pairwise_distance,
-        "single_to_multiple_distance": euclidean_from_single_to_multiple_distance,
-        "multiple_to_multiple_distance": euclidean_from_multiple_to_multiple_distance,
+        "single_to_multiple_distance": euclidean_pairwise_distance,
+        "multiple_to_multiple_distance": euclidean_pairwise_distance,
     },
     {
         "name": "squared",
