+++ setup.cfg
@@ -11,15 +11,15 @@ addopts =
     --ignore build_tools
     --ignore examples
     --ignore docs
-    --doctest-modules
-    --durations 10
-    --timeout 600
-    --cov aeon
-    --cov-report xml
-    --cov-report html
-    --showlocals
-    --matrixdesign True
-    -n auto
+;    --doctest-modules
+;    --durations 10
+;    --timeout 600
+;    --cov aeon
+;    --cov-report xml
+;    --cov-report html
+;    --showlocals
+;    --matrixdesign True
+;    -n auto
 filterwarnings =
     ignore::UserWarning
     ignore:numpy.dtype size changed
